Implement error handling on createOrder routes

+++ services/CreateOrder.service.js
@@ -1,4 +1,9 @@
 import CreateOrderModel from "../models/CreateOrder.model.js"
+import {
+  BadRequestError,
+  NotFoundError,
+  ValidationError,
+} from "../utils/customErrorHandler.js"
 
 export const getAllItems = async (req, res) => {
   try {
@@ -12,7 +17,13 @@ export const getAllItems = async (req, res) => {
 
 export const getItemsByUserId = async (req, res) => {
   try {
+    if (!req.params.id) {
+      throw new BadRequestError("Not mentioned user id on API.")
+    }
     const allItems = await CreateOrderModel.find({ userId: req.params.id })
+    if (allItems.length === 0) {
+      throw new NotFoundError("No create Order found.")
+    }
     res.status(200)
     res.json(allItems)
   } catch (error) {
@@ -22,19 +33,39 @@ export const getItemsByUserId = async (req, res) => {
 
 export const saveNewItem = async (req, res) => {
   try {
+    if (!req.body) {
+      throw new BadRequestError("Request body is missing.")
+    }
     const NewItem = new CreateOrderModel(req.body)
     await NewItem.save()
     res.status(200)
     res.json(NewItem)
   } catch (error) {
-    throw error
+    if (error.name === "ValidationError") {
+      throw new ValidationError(error.message)
+    } else if (error.code === 11000) {
+      const field = Object.keys(error.keyPattern)[0]
+      throw new ValidationError(`${field} must be unique`)
+    } else {
+      throw error
+    }
   }
 }
 
 export const findCreateOrderByUserIdAndUpdate = async (req, res) => {
   try {
     const id = req.params.id
+    if (!id) {
+      throw new BadRequestError("User id is missing.")
+    }
+    const CreateOrder = await CreateOrderModel.find({ userId: id })
+    if (CreateOrder.length === 0) {
+      throw new NotFoundError("No create order found.")
+    }
     const dataToUpdate = req.body
+    if (!dataToUpdate) {
+      throw new BadRequestError("Request body is missing.")
+    }
     const createOrder = await CreateOrderModel.findOneAndUpdate(
       { userId: id },
       dataToUpdate,
@@ -51,6 +82,13 @@ export const findCreateOrderByUserIdAndUpdate = async (req, res) => {
 
 export const findByUserIdAndDelete = async (req, res) => {
   try {
+    if (!req.params.id) {
+      throw new BadRequestError("User id is missing.")
+    }
+    const CreateOrder = await CreateOrderModel.find({ userId: req.params.id })
+    if (CreateOrder.length === 0) {
+      throw new NotFoundError("No create order found.")
+    }
     const item = await CreateOrderModel.findOneAndDelete({
       userId: req.params.id,
     })
